refactor(search): extract /api/v1/db/search into src/api/search.js
Pure code-organisation cleanup. No behaviour changes — same routes,
same SQL, same response shapes, same algorithm dispatch.

Why: src/api/db.js had grown to ~700 LOC since PR2 added the FTS5
helpers and PR3 added the algorithm-dispatch route. The search code
is now bigger than the rest of /api/v1/db/* put together but only
shares one helper (libraryFilter) with it.

Moved to src/api/search.js:
  - shape* callbacks (4 exported pure functions)
  - logFts5UnavailableOnce + module-scoped latch
  - per-category LIKE row builders (4)
  - runLikeSearch (exported)
  - per-category FTS5 row builders (4)
  - runFtsSearch (exported)
  - /api/v1/db/search route + Joi schema

Stayed in src/api/db.js:
  - renderMetadataObj, libraryFilter, trackQuery, pullMetaData
    (the four cross-cutting helpers other route modules and
    src/util/file-explorer.js consume)
  - All other /api/v1/db/* route handlers

Stayed in src/util/search-query.js:
  - parseSearchQuery, escapeFts, buildFtsExpression
    (DB-free utility functions; also consumed by
    src/api/subsonic/handlers.js — keeping them out of the api/
    tree avoids a cross-import between two api modules)

Wired src/server.js to call searchApi.setup(mstream) right after
dbApi.setup(mstream) so the registration ordering matches the
auth-required group around it.

Tests: 535/535 pass on full suite — same count as before this commit.
