+++ src/server.js
@@ -10,6 +10,7 @@ import https from 'https';
 import { createRequire } from 'module';
 
 import * as dbApi from './api/db.js';
+import * as searchApi from './api/search.js';
 import * as playlistApi from './api/playlist.js';
 import * as authApi from './api/auth.js';
 import * as fileExplorerApi from './api/file-explorer.js';
@@ -251,6 +252,7 @@ export async function serveIt(configFile) {
 
   adminApi.setup(mstream);
   dbApi.setup(mstream);
+  searchApi.setup(mstream);
   playlistApi.setup(mstream);
   downloadApi.setup(mstream);
   fileExplorerApi.setup(mstream);
